+++ tests/TestCase.php
@@ -31,8 +31,9 @@ class TestCase extends OrchestraTestCase
     protected function getEnvironmentSetUp($app)
     {
         $app->setBasePath(__DIR__.'/../');
-        $app['config']->set('view-export.chroot', $app->basePath().'/vendor/dompdf/dompdf');
-        chmod(config('view-export.chroot'), 0755);
+        $app['config']->set('view-export.pdf.chroot', $app->basePath().'/vendor/dompdf/dompdf');
+        $app['config']->set('view-export.pdf.font_cache', $app->storagePath().'/fonts');
+        chmod(config('view-export.pdf.chroot'), 0755);
     }
 
     /**